Remove strdup from InfixParser
InfixParser was duplicating strings to be passed to SyntaxTree
without freeing them afterwards, expecting that SyntaxTree would
own the duplicate pointer and free it later. SyntaxTree was
duplicating strings it received as input so that the caller
didn't have to diplicate strings, and could use constants as
values. This lead to a memory leak.

Bug reported by Bao Hexing.